chore: merge conftest files to suppress ambiguities
